print 50 fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,14 +7,14 @@
  */
 int main(void)
 {
-	int n1 = 1, n2 = 2;
-	int i, n3;
+	long int n1 = 1, n2 = 2;
+	long int i, n3;
 
-	printf("%d, %d, ", n1, n2);
+	printf("%ld, %ld, ", n1, n2);
 	for (i = 3; i < 51; i++)
 	{
 		n3 = n1 + n2;
-		printf("%d, ", n3);
+		printf("%ld, ", n3);
 		n1 = n2;
 		n2 = n3;
 	}
